refactor comment delete
